+++ src/subcommand/dist_build.rs
@@ -1,6 +1,6 @@
 use std::any::Any;
 
-use crate::{config::Config, Result, Run};
+use crate::{config::Config, Result, Run, SubcommandRun};
 
 /// Arguments definition of the `dist-build` subcommand.
 #[cfg_attr(doc, doc = include_str!("../../doc/cargo-xtask-dist-build.md"))]
@@ -47,6 +47,10 @@ pub struct DistBuild {
 impl Run for DistBuild {
     fn run(&self, config: &Config) -> Result<()> {
         self.run(config)
+    }
+
+    fn to_subcommands(&self) -> Option<SubcommandRun> {
+        None
     }
 
     fn into_any(self: Box<Self>) -> Box<dyn Any> {