+++ src/subcommand/dist_build_license.rs
@@ -2,7 +2,7 @@ use std::any::Any;
 
 use eyre::eyre;
 
-use crate::{config::Config, fs::ToRelative, Result, Run};
+use crate::{config::Config, fs::ToRelative, Result, Run, SubcommandRun};
 
 /// Arguments definition of the `dist-build-license` subcommand.
 #[cfg_attr(doc, doc = include_str!("../../doc/cargo-xtask-dist-build-license.md"))]
@@ -13,6 +13,10 @@ pub struct DistBuildLicense {}
 impl Run for DistBuildLicense {
     fn run(&self, config: &Config) -> Result<()> {
         self.run(config)
+    }
+
+    fn to_subcommands(&self) -> Option<SubcommandRun> {
+        None
     }
 
     fn into_any(self: Box<Self>) -> Box<dyn Any> {