+++ xtask/src/main.rs
@@ -1,6 +1,8 @@
 use std::any::Any;
 
-use cli_xtask::{clap, config::Config, subcommand::Subcommand as Predefined, Result, Run, Xtask};
+use cli_xtask::{
+    clap, config::Config, subcommand::Subcommand as Predefined, Result, Run, SubcommandRun, Xtask,
+};
 
 mod lint;
 mod lint_doc;
@@ -31,6 +33,10 @@ impl Run for Subcommand {
             Self::XtaskTest(args) => args.run(config)?,
         }
         Ok(())
+    }
+
+    fn to_subcommands(&self) -> Option<SubcommandRun> {
+        None
     }
 
     fn into_any(self: Box<Self>) -> Box<dyn Any> {