+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Fixed
+
+* Removed a wrong default implementation of `Run::to_subcommands()`
+
 ## [0.3.0] - 2022-09-16
 
 ### Added