+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+## [0.5.0] - 2022-09-16
+
 ### Added
 
 * Add `Subcommand::selected()` method
@@ -66,7 +68,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * First release
 
 <!-- next-url -->
-[Unreleased]: https://github.com/gifnksm/cli-xtask/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/gifnksm/cli-xtask/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/gifnksm/cli-xtask/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/gifnksm/cli-xtask/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/gifnksm/cli-xtask/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/gifnksm/cli-xtask/compare/v0.1.4...v0.2.0